remove unit tests and cover a few more edge cases with integ tests
I've decided that "intent-based" testing is better than "effect-based". As a result, all tests are now integration tests -- they start with the initialization of a store, perform some action on it, and then check the result. The tests are slightly more resillient this way and are easier to read.

In addition, we won't need to expose internal behaviour nearly as much. (e.g. select.set returning an action instead of, say, the value we set it to).